Struct aloft::StationForecast

pub struct StationForecast {
    // some fields omitted
}

Weather data from one weather station at all altitudes

Methods

impl StationForecast

fn new(station: &str, winds: Vec<Wind>) -> StationForecast

fn wind_at_altitude(&self, altitude: u32) -> Option<&Wind>

Search through the wind data from this station, return the wind data from the given altitude if it exists, otherwise return None.