Struct aloft::WindsAloftForecast

pub struct WindsAloftForecast {
    pub forecasts: Vec<StationForecast>,
    // some fields omitted
}

All information associated with winds aloft forecast

Fields

forecasts

The list of StationForecast objects that are held in this report.

Methods

impl WindsAloftForecast

fn new() -> WindsAloftForecast

fn refresh(&mut self) -> Option<()>

Re-downloads the forecast from NOAA. Returns None if there is a problem with the download, or Some(()) if the operation completes successfully.

fn needs_refresh(&self) -> bool

Returns true if the data stored in this forecast is out of date, false otherwise. The winds aloft forecast is published at 00Z, 06Z, 12Z and 18Z

fn get_station_forecast(&self, station: &str) -> Option<&StationForecast>

Get the forecast for a specific station out of this complete forecast.